api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
requests==2.32.3
pyjwt==2.9.0
python-dateutil==2.9.0
